refactor: split entries route handlers
